requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3
python-dotenv==1.0.0
pydantic==2.5.0
streamlit==1.29.0
python-telegram-bot==20.7
requests==2.31.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
pandas==2.1.3
plotly==5.18.0